chore rename uploadImg mutation in hooks
